requirements.txt
torch==1.7.1
torchvision==0.8.2
facenet-pytorch==2.5.1
face_recognition
cmake
dlib
numpy
Pillow